+++ src/components/Configure.mdx
@@ -1,0 +1,12 @@
+import { Meta } from '@storybook/blocks';
+
+<Meta title='Configure your project' />
+
+<div className="sb-container">
+  <div className='sb-section-title'>
+    # Configure your project
+
+    Because Storybook works separately from your app, you'll need to configure it for your specific stack and setup. Below, explore guides for configuring Storybook with popular frameworks and tools. If you get stuck, learn how you can ask for help from our community.
+
+  </div>
+</div>
