+++ src/components/button/index.stories.ts
@@ -1,6 +1,6 @@
-import type { Meta, StoryObj } from '@storybook/react';
+import { Button } from '@/components';
 
-import { Button } from './Button';
+import type { Meta, StoryObj } from '@storybook/react';
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
