+++ src/components/header/index.stories.ts
@@ -1,6 +1,6 @@
-import type { Meta, StoryObj } from '@storybook/react';
+import { Header } from '@/components';
 
-import { Header } from './Header';
+import type { Meta, StoryObj } from '@storybook/react';
 
 const meta = {
   title: 'Example/Header',
