+++ src/components/index.ts
@@ -1,0 +1,4 @@
+import { Button } from './button';
+import { Header } from './header';
+
+export { Button, Header };
